inline GitHubWebhookVerifier into the handler; add no-single-file-folder rule
The verifier was a single static method behind two private helpers, used in
exactly one place. Inlining drops the class + the security/ subpackage
that held only it. The HMAC-SHA256 + constant-time compare logic lives
directly in GitHubWebhookHandler now.

Add a project-level rule to CLAUDE.md to capture the principle: don't
create a package or folder to hold a single file.
