+++ CLAUDE.md
@@ -49,5 +49,7 @@ When create code to store data call it a Service not a Store. If the Service nee
 
 Configuration follows the same split: `api/config/` contains `@ConfigurationProperties` classes and settings POJOs meant to be configured by users, while `internal/config/` contains Spring `@Configuration` classes that wire beans internally. This applies to all modules.
 
+Don't create a new package or folder to hold a single file. Single-file folders just spread related code across the tree without aiding discoverability. Place the file in the nearest existing package that fits, or inline its contents into its only caller. A new subpackage is justified once there are at least two or three related files that genuinely belong together.
+
 ## Properties
 Properties should never be created for something that will not need to be configured differently in different environments. i.e. Kinotic Cloud dev vs Kinotic Cloud prod. In the case of a route or something that will be the same for multiple environments, create a constant.
