+++ kinotic-github/src/main/java/org/kinotic/github/internal/rest/GitHubWebhookHandler.java
@@ -13,8 +13,12 @@ import org.kinotic.github.api.config.KinoticGithubProperties;
 import org.kinotic.github.api.model.GitHubWebhookEvent;
 import org.kinotic.github.api.rest.GithubConstants;
 import org.kinotic.github.api.services.GitHubWebhookEventService;
-import org.kinotic.github.internal.security.GitHubWebhookVerifier;
 import org.springframework.stereotype.Component;
+
+import javax.crypto.Mac;
+import javax.crypto.spec.SecretKeySpec;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
 
 /**
  * {@code POST /api/github/webhook}: HMAC-verifies the delivery, parses the JSON, and
@@ -23,8 +27,9 @@ import org.springframework.stereotype.Component;
  * The {@link BodyHandler} is route-scoped (not global) with a 25 MiB cap matching
  * GitHub's documented webhook ceiling; oversized payloads are rejected with 413
  * before HMAC ever runs. HMAC verification is computed against the raw bytes
- * <strong>before</strong> any JSON decode — see {@link GitHubWebhookVerifier} —
- * since whitespace differences would invalidate the signature.
+ * <strong>before</strong> any JSON decode — whitespace differences would invalidate
+ * the signature. Compare is constant-time so the expected digest doesn't leak
+ * through timing.
  * <p>
  * Always returns 204 quickly so GitHub doesn't redeliver. Any internal failure is
  * logged and dropped by the dispatch service.
@@ -37,6 +42,9 @@ public class GitHubWebhookHandler {
     private static final String HEADER_EVENT = "X-GitHub-Event";
     private static final String HEADER_DELIVERY = "X-GitHub-Delivery";
     private static final String HEADER_SIGNATURE = "X-Hub-Signature-256";
+    private static final String SIGNATURE_PREFIX = "sha256=";
+    private static final String HMAC_ALGO = "HmacSHA256";
+    private static final char[] HEX = "0123456789abcdef".toCharArray();
 
     /** GitHub's documented webhook ceiling — anything larger is rejected with 413. */
     private static final long WEBHOOK_BODY_LIMIT_BYTES = 25L * 1024 * 1024;
@@ -71,7 +79,7 @@ public class GitHubWebhookHandler {
         byte[] bodyBytes = body.getBytes();
         String secret = properties.getGithub().getWebhookSecret();
         Future<Boolean> verify = vertx.executeBlocking(
-                () -> GitHubWebhookVerifier.verify(bodyBytes, secret, signature),
+                () -> verifySignature(bodyBytes, secret, signature),
                 false);
         verify.onComplete(ar -> {
             if (ar.failed() || !Boolean.TRUE.equals(ar.result())) {
@@ -98,6 +106,39 @@ public class GitHubWebhookHandler {
                 }
             });
         });
+    }
+
+    /**
+     * HMAC-SHA256 the raw body with the App's webhook secret and compare (constant-time)
+     * against the {@code sha256=<hex>} signature GitHub sent in the header.
+     */
+    private static boolean verifySignature(byte[] rawBody, String secret, String headerSignature) {
+        if (rawBody == null || secret == null || secret.isEmpty()
+                || headerSignature == null || !headerSignature.startsWith(SIGNATURE_PREFIX)) {
+            return false;
+        }
+        String expectedHex;
+        try {
+            Mac mac = Mac.getInstance(HMAC_ALGO);
+            mac.init(new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), HMAC_ALGO));
+            expectedHex = toHex(mac.doFinal(rawBody));
+        } catch (Exception e) {
+            return false;
+        }
+        String received = headerSignature.substring(SIGNATURE_PREFIX.length());
+        return MessageDigest.isEqual(
+                expectedHex.getBytes(StandardCharsets.US_ASCII),
+                received.getBytes(StandardCharsets.US_ASCII));
+    }
+
+    private static String toHex(byte[] bytes) {
+        char[] out = new char[bytes.length * 2];
+        for (int i = 0; i < bytes.length; i++) {
+            int v = bytes[i] & 0xFF;
+            out[i * 2] = HEX[v >>> 4];
+            out[i * 2 + 1] = HEX[v & 0x0F];
+        }
+        return new String(out);
     }
 
     private static GitHubWebhookEvent buildEvent(String eventType,
